+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [v0.7.1]
 ### Fixed
 - Spacing units calculations internally. Mobile layouts improved.
 
@@ -93,6 +95,7 @@ All notable changes to this project will be documented in this file. The format 
 ### Other
 - vibe coded
 
+[v0.7.1]: https://github.com/off-court-creations/valet/releases/tag/v0.7.1
 [v0.7.0]: https://github.com/off-court-creations/valet/releases/tag/v0.7.0
 [v0.6.1]: https://github.com/off-court-creations/valet/releases/tag/v0.6.1
 [v0.6.0]: https://github.com/off-court-creations/valet/releases/tag/v0.6.0